TC_09.006.03 | Manage Jenkins >System log >Clear the custom log recorder
